+++ app/Http/Controllers/Admin/Track/Api/DestroyController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\Admin\Track\Api;
 
 use App\Http\Controllers\Admin\Track\BaseController;
+use App\Models\AuditTrack;
 use App\Models\Track;
 use App\Models\User;
 use OwenIt\Auditing\Models\Audit;
@@ -11,23 +12,26 @@ class DestroyController extends BaseController
 {
     public function __invoke(Track $track, User $user)
     {
-        if (($path = $track->artwork_url)) {
-            $path = substr($path, 0, strpos($path, basename($path)));
-            $path = pathToServer() . substr($path, strpos($path, "images"));
-            $path = preg_replace('/artwork\//', '', $path);
-            $this->service->delete($path);
+        $moderator = 3;
+
+        if ($user->role !== $moderator) {
+            if (($path = $track->artwork_url)) {
+                $path = substr($path, 0, strpos($path, basename($path)));
+                $path = pathToServer() . substr($path, strpos($path, "images"));
+                $path = preg_replace('/artwork\//', '', $path);
+                $this->service->delete($path);
+            }
+
+            $track->artists()->detach();
+            $track->genres()->detach();
+            $track->album()->detach();
         }
-
-
-        $track->artists()->detach();
-        $track->genres()->detach();
-        $track->album()->detach();
 
         $track->delete();
 
         $audit =  Audit::latest()->first();
 
-        // $audit->update(['user_type' => 'App\Model\User', 'user_id' => $user->id]);
+        $audit->update(['user_type' => 'App\Model\User', 'user_id' => $user->id]);
 
         return response([]);
     }
